Patch for non-interactive text sources on startup
On startup and text creation there was an issue where the slideshow
would not be interactable. For freetype2 this was due to the new text
source returning a width and height of 0. Subsequent updates after the
first update would fix the issue. Some code to dynamically keep track of
the width and height when available was added to the video render
callback. It is a janky fix until the freetype2 issue of not properly
setting the width and height can be fixed