Update to support OBS v28.0.0 and v29.0.0 (#53)
* Updates from new plugin template

* Removed old clang github action

* Allowed building on all commits

* buildspec updates

* Fixed linux build

* Fix windows pthread linking

* Fix duplicate target_link_libraries calls

* Include pthread.h

* Try direct w32-pthread linking

* Fixed windows build

* Fixed CMakeLists.txt formatting

* Fixed installation target for Ubuntu 22.04.1

* Refactored manual refresh on frontend startup

Issue that was present on gdiplus now showed up in freetype2. Janky existing
fix was refactored to be added to text-freetype2

* Re-added cmake installation call for macos

cmake on linux will perform the local install after building while macos
will not. This re-adds the installation call for macos so the packaging
script does not fail

* Fixed styling error

* Pulled in missing commits from obs-studio main repo

* Patch for non-interactive text sources on startup

On startup and text creation there was an issue where the slideshow
would not be interactable. For freetype2 this was due to the new text
source returning a width and height of 0. Subsequent updates after the
first update would fix the issue. Some code to dynamically keep track of
the width and height when available was added to the video render
callback. It is a janky fix until the freetype2 issue of not properly
setting the width and height can be fixed

* Switched to unordered_set for size patch over DARRAY

* Cleared out of date sources and updated comment

* Single file delimiter parsing fixes (#49)

* Complete rewrite of delim parsing logic

Complete rewrite of the delim parsing logic with a custom multi
character strtok implmentation to fix a lot of the lingering
delim parsing bugs due to strtok limitations

* Refactor and rewrite delim file parsing

Overhaul and refactor delimeter file reading and parsing

* Fixed windows build errors

* Fix for multi file support

Fix for multi file support which was broken by the empty
width/height patch

* Attempt to fix Ubuntu locale install path (#50)